browser-use-backend/requirements.txt
fastapi~=0.116.0
uvicorn[standard]
python-dotenv~=1.1.1
browser-use==0.5.11
openai==1.99.2
pillow
websockets
docker
requests==2.32.3
langchain-openai==0.3.27
pydantic==2.11.7
pydantic-settings>=2.5.2
motor==3.1.1
pymongo==4.3.3
authlib
starlette~=0.46.2
azure-storage-blob
capsolver
psutil~=7.0.0
azure-identity
azure-keyvault-secrets
itsdangerous>=2.1.2